alt_e2eshark/requirements.txt
#install nightly build of torch_mlir, if on Linux (no macOS or Windows nightly builds)
-f https://github.com/llvm/torch-mlir-release/releases/expanded_assets/dev-wheels
torch-mlir ; sys_platform == "linux"
# install nightly build of iree-compiler and iree-runtime
iree-compiler -f https://iree.dev/pip-release-links.html
iree-runtime -f https://iree.dev/pip-release-links.html
tabulate
simplejson
ml_dtypes
onnx
onnxruntime
transformers
huggingface-hub
sentencepiece
accelerate
auto-gptq
optimum
azure-storage-blob